+++ Source/ColorAction.cpp
@@ -35,16 +35,16 @@ bool ColorAction::initWithDuration(float duration, SpriteColor* target, ax::Colo
 void ColorAction::startWithTarget(ax::Node* target)
 {
 	ActionInterval::startWithTarget(target);
-	_deltar = _to.r - _from.r;
-	_deltag = _to.g - _from.g;
-	_deltab = _to.b - _from.b;
+	_deltar = static_cast<float>(_to.r - _from.r);
+	_deltag = static_cast<float>(_to.g - _from.g);
+	_deltab = static_cast<float>(_to.b - _from.b);
 	_deltaa = _ato - _afrom;
 }
 
 void ColorAction::update(float dt)
 {
-	_target->_color.r = _to.r - _deltar * (1 - dt);
-	_target->_color.g = _to.g - _deltag * (1 - dt);
-	_target->_color.b = _to.b - _deltab * (1 - dt);
+	_target->_color.r = static_cast<uint8_t>(_to.r - _deltar * (1 - dt));
+	_target->_color.g = static_cast<uint8_t>(_to.g - _deltag * (1 - dt));
+	_target->_color.b = static_cast<uint8_t>(_to.b - _deltab * (1 - dt));
 	if(_ato >= 0) _target->_opacity = (_ato * 255.f) - _deltaa * (1 - dt);
 }
